% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{as_se}
\alias{as_se}
\title{Coercion to \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}
\usage{
as_se(npx)
}
\arguments{
\item{npx}{\code{\link[tibble]{tibble}} object from \code{\link{read_npx}}.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.
}
\description{
Coercion to \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.
}
\examples{
npxFn <- system.file("extdata", "OlinkAnalyze_extdata", "npx_data2.xlsx",
                     package = "OlinkR")
metaFn <- system.file("extdata", "OlinkAnalyze_extdata",
                      "npx_data2_meta.xlsx", package = "OlinkR")
npx <- read_npx(npxFn, metaFn)
se <- as_se(npx)
}
\author{
Ge Tan
}
