% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{.renamePanels}
\alias{.renamePanels}
\title{Rename panel name}
\usage{
.renamePanels(panels)
}
\arguments{
\item{panels}{\code{character}(n): the panel names}
}
\description{
Rename panel name to be consistent between different years/batches.
}
\author{
Ge Tan
}
