% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{list_panels}
\alias{list_panels}
\title{List panels from NPX files}
\usage{
list_panels(npxFn)
}
\arguments{
\item{npxFn}{Path to NPX files}
}
\value{
A \code{character}(n) object.
}
\description{
Read through the given NPX files and return all panels inside
}
\examples{
npxFn <- system.file("extdata", c(
  "2019-01-02_20181145_Akdis_Nasal_NPX_LOD.xlsx"
),
package = "OlinkR"
)
list_panels(npxFn)
}
\author{
Ge Tan
}
