% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Olink_boxplot.R
\name{olink_boxplot}
\alias{olink_boxplot}
\title{Function which plots boxplots of a selected variable}
\usage{
olink_boxplot(df, variable, olinkid_list,
  number_of_proteins_per_plot = 6)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID (unique), UniProt and a grouping variable.}

\item{variable}{A character value indiciating which column to use as the x-axis grouping variable}

\item{olinkid_list}{Character vector indicating for which OlinkID's to create figures}

\item{number_of_proteins_per_plot}{Number boxplots to include in the array of boxplots. (defaults 6 plots per figure)}
}
\value{
A list of objects of class "ggplot"
}
\description{
Creates an array of boxplots with grouping variable on the x-axis and NPX values on the y-axis for a given list of proteins (by OlinkID).
}
\examples{
\donttest{anova_results <- anova_olink(df, variable = "Var1")
significant_assays <- anova_results \%>\%
filter(Threshold = 'Significant') \%>\%
pull(OlinkID)
olink_boxplot(df, variable = "Var1", olinkid_list = significant_assay, number_of_proteins_per_plot = 3)}
}
