% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{olink_limma}
\alias{olink_limma}
\title{limma analysis for Olink data}
\usage{
olink_limma(se, factorCol, contrasts, blocking = NULL)
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}

\item{factorCol}{\code{character}(1): The metadata column in \code{se} that defines the grouping.}

\item{contrasts}{\code{character}(1): A character string which can be parsed to expressions, specifying contrasts.}

\item{blocking}{\code{character}(1): The metadata column in \code{se} which serves as a blocking factor.}
}
\value{
A \code{tibble} object of results from limma analysis
}
\description{
Given a \code{SummarizedExperiment} object, perform two groups comparison with \code{limma} package.
}
\examples{
npxFn <- system.file("extdata", c("20200507_Inflammation_NPX_1.xlsx",
                                  "20200625_Inflammation_NPX_2.xlsx"),
                     package="OlinkR")
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package="OlinkR")
se <- readNPX(npxFn, metaFn)$SummarizedExperiment
tb <- olink_limma(se, factorCol="condition_Factor",
                  contrasts="Glucose.10mM.Vehicle - Vehicle.Vehicle",
                  blocking="Donor_Factor")
}
\author{
Ge Tan
}
