% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mixed_model.R
\name{olink_lmer_posthoc}
\alias{olink_lmer_posthoc}
\title{Function which perform a linear mixed model posthoc per significant protein}
\usage{
olink_lmer_posthoc(df, var1, var2, ID, lmer_result, variable = 3,
  mean.return = F)
}
\arguments{
\item{df}{NPX data frame in long format with at least protein name (Assay), OlinkID, UniProt, 1-2 variables with at least 2 levels and subject ID.}

\item{var1}{Character value indicating the name of the column to be used as the first variable in the LME model}

\item{var2}{Character value indicating the name of the column to be used as the second variable in the LME model}

\item{ID}{Character value indicating the column of subject ID's. To be used as a random effect in the LME model.}

\item{lmer_result}{Data frame containing the results of fitting the same model to the NPX data using the olink_lmer_main function.}

\item{variable}{Integer. Which variable on which the posthoc is performed. Must be either 1,2,3. 1 = var1, 2 = var2, 3 = interaction term.}

\item{mean.return}{Logical. Should the estimated marginal means for each group be returned (default) or the difference between them.}
}
\value{
A data frame containing the results of the pairwise comparisons between given variable levels for proteins found to be statistically significant by olink_lmer_main.
}
\description{
Performs post-hoc analysis for proteins found to be statistically significant by the olink_lmer_main function. Differences in estimated marginal means are calculated for all pairwise levels of a given variable. Degrees of freedom estimated using Satterthwaite's approximation.
}
\examples{
\donttest{lme_results <- olink_lmer_main(df=npx.data, var1 = "Visit", var2 = "Treatment", ID = "SubjectID")
results_lmer_posthoc <- olink_lmer_posthoc(df=npx.data, var1 = "Visit", var2 = "Treatment", ID = "SubjectID", lmer_result=lme_results, variable = 1, mean.return = T)}
}
