% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{olink_pca}
\alias{olink_pca}
\title{PCA analysis of olink data}
\usage{
olink_pca(se, colby = NULL, shape = NULL, metavars = NULL, removeVar = 0.1)
}
\arguments{
\item{se}{An \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}

\item{colby}{\code{character}(1): The metadata column in \code{se} to use for colour.}

\item{shape}{\code{character}(1): The metadata column in \code{se} to use for shape.}

\item{metavars}{\code{character}(n): The metedata columns with numeric values in \code{se} to correlate with principal components.}

\item{removeVar}{\code{numeric}(1): Remove this fraction of variables based on low variance. DEFAULT = 0.1. OPTIONAL.}
}
\value{
A list of \code{ggplot} and \code{tibble} objects.
}
\description{
Given a \code{SummarizedExperiment} object, perform the PCA analysis implemented
in Bioconductor package \code{PCAtools}
}
\examples{
npxFn <- system.file("extdata", c("20200507_Inflammation_NPX_1.xlsx",
                                  "20200625_Inflammation_NPX_2.xlsx"),
                     package="OlinkR")
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package="OlinkR")
se <- read_npx(npxFn, metaFn)$SummarizedExperiment
ans <- olink_pca(se, colby="condition_Factor", shape="Donor_Factor", metavars="Weight_Numeric")
}
\author{
Ge Tan
}
