% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{olink_volcano}
\alias{olink_volcano}
\title{Volcano plot of Olink results}
\usage{
olink_volcano(tb, p.value = 0.05, log2FC = 0, olinkIds = NULL)
}
\arguments{
\item{tb}{A \code{tibble} object from \code{olink_limma} function.}

\item{p.value}{\code{numeric}(1): The p.value cutoff for significance level.}

\item{log2FC}{\code{numeric}(1): The log2FC cutoff for significance level.}

\item{olinkIds}{\code{character}(n): OlinkIDs to label on the volcano plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Given the \code{tibble} object from \code{olink_limma}, plot a volcano plot.
The log2FC is on the x-axis and -log10(P.value) is on y-axis.
By default, significant proteins are labeled on the volcano plot.
}
\examples{
npxFn <- system.file("extdata", c("20200507_Inflammation_NPX_1.xlsx",
                                  "20200625_Inflammation_NPX_2.xlsx"),
                     package="OlinkR")
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package="OlinkR")
se <- read_npx(npxFn, metaFn)$SummarizedExperiment
tb <- olink_limma(se, factorCol="condition_Factor",
                  contrasts="Glucose.10mM.Vehicle - Vehicle.Vehicle",
                  blocking="Donor_Factor")
olink_volcano(tb)
}
\author{
Ge Tan
}
