% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{overview_heatmap}
\alias{overview_heatmap}
\title{A overview of NPX value in heatmap}
\usage{
overview_heatmap(
  se,
  scale = c("none", "row", "column"),
  cluster_samples = FALSE,
  cluster_features = FALSE,
  ...
)
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}

\item{scale}{\code{character}(1): how to scale (z-score) the data. "none": no scaling.
"sample": scale across the samples.
"feature": scale across the features (proteins).}

\item{cluster_samples}{\code{logical}(1): do the samples clustering.}

\item{cluster_features}{\code{logical}(1): do the samples clustering.}

\item{...}{graphical parameter passed to \code{pheatmap}.}
}
\value{
A \code{pheatmap} object.
}
\description{
Plot the NPX value in a heatmap to spot any potential clustering and outliers
}
\examples{
npxFn <- system.file("extdata",
                     c("20200507_Inflammation_NPX_1.xlsx",
                       "20200625_Inflammation_NPX_2.xlsx"),
                     package = "OlinkR")
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package = "OlinkR")
se <- read_npx(npxFn, metaFn)$SummarizedExperiment
overview_heatmap(se)
}
\author{
Ge Tan
}
