% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read NPX data.R
\name{read_NPX}
\alias{read_NPX}
\title{Function to read NPX data into long format}
\usage{
read_NPX(filename)
}
\arguments{
\item{filename}{filename of NPX file exported from NPX manger.}
}
\value{
A tibble in long format.
}
\description{
Imports an NPX file exported from NPX manager. Must be an MS Excel file.
}
\examples{
\donttest{read_NPX("~/NPX data.xlsx")}
}
\keyword{NPX}
