% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webgestalt.R
\name{webgestalt_prep}
\alias{webgestalt_prep}
\title{Produce files for webgestalt website}
\usage{
webgestalt_prep(x, pvalue = 0.01, log2FC = 0, dir = ".")
}
\arguments{
\item{x}{A \code{tibble} object from \code{olink_limma} function.}

\item{pvalue}{The p-value to filter for significant proteins. The default is 0.01.}

\item{log2FC}{The log2 fold change cut-off to filter for significant proteins. The default is 0, i.e. no filtering.}

\item{dir}{The output folder. The default is current working directory.}
}
\value{
An invisible character object with exported file names.
}
\description{
Given a \code{tibble} object of the results returned from \code{olink_limma} function, generate the files for upload to webgestalt website.
}
\examples{
npxFn <- system.file("extdata", c(
  "20200507_Inflammation_NPX_1.xlsx",
  "20200625_Inflammation_NPX_2.xlsx"
),
package = "OlinkR"
)
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package = "OlinkR")
se <- read_npx(npxFn, metaFn)$SummarizedExperiment
tb <- olink_limma(se,
  factorCol = "condition_Factor",
  contrasts = "Glucose.10mM.Vehicle - Vehicle.Vehicle",
  blocking = "Donor_Factor"
)
files <- webgestalt_prep(se)
file.remove(files)
}
\author{
Ge Tan
}
