% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{olink_boxplot}
\alias{olink_boxplot}
\title{Boxplot for Olink Data}
\usage{
olink_boxplot(se, features, x = NULL)
}
\arguments{
\item{se}{a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.}

\item{features}{\code{character}(n): the protein names to plot.}

\item{x}{\code{character}(1): the grouping colname in \code{colData(se)}.}
}
\value{
A list of \code{ggplot} objects
}
\description{
Generate a list of boxplot of NPX vs. grouping variable.
}
\examples{
npxFn <- system.file("extdata", c(
  "20200507_Inflammation_NPX_1.xlsx",
  "20200625_Inflammation_NPX_2.xlsx"
),
package = "OlinkR"
)
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package = "OlinkR")
se <- read_npx(npxFn, metaFn)$SummarizedExperiment
features <- c("IL8", "MCP-3")
pList <- olink_boxplot(se, features, x = "condition_Factor")
}
\author{
Ge Tan
}
