% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{read_npx}
\alias{read_npx}
\title{Read NPX and metadata file together}
\usage{
read_npx(npxFn, metaFn, panel = NULL)
}
\arguments{
\item{npxFn}{Path to NPX files}

\item{metaFn}{Path to metadata excel file}

\item{panel}{\code{character}(1): the panel to load.
By default, the first panel in the data will be loaded.}
}
\value{
A \code{\link[tibble]{tibble}} in long format with following columns:
\describe{
\item{SampleID}{The sample name}
\item{OlinkID}{The internal Olink IDs}
\item{UniProt}{The UniProt IDs}
\item{Assay}{The protein name}
\item{MissingFreq}{The percentage of samples that have intensities lower than LOD}
\item{Panel}{The panel name}
\item{Panel_Version}{The panel version}
\item{PlateID}{The unique plate ID}
\item{QC_Warning}{If the splot passes QC}
\item{LOD}{The limit of detection}
\item{NPX}{NPX value}
\item{Normalization}{The intr-plate normalization method}
}
}
\description{
Read one or more NPX files from Olink NPX manager and
one metadata file in excel format.
}
\details{
The metadata excel file shall have column \dQuote{SampleID} and other metadata column that ends with \dQuote{_Factor} (mandatory )or \dQuote{_Numeric} (optional).
If some samples from NPX don't exist in metadata file, it gives a warning.
Only the samples in metadata file are kept.
If some samples from metadata file don't exists in NPX file, it gives an error.
}
\examples{
npxFn <- system.file("extdata",
                     c("20200507_Inflammation_NPX_1.xlsx",
                       "20200625_Inflammation_NPX_2.xlsx"),
                     package = "OlinkR")
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package = "OlinkR")
npx <- read_npx(npxFn, metaFn)

npxFn <- system.file("extdata", "OlinkAnalyze_extdata", "npx_data2.xlsx",
                     package = "OlinkR")
metaFn <- system.file("extdata", "OlinkAnalyze_extdata" ,
                      "npx_data2_meta.xlsx", package = "OlinkR")
npx <- read_npx(npxFn, metaFn)
}
\author{
Ge Tan
}
