% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npx.R
\name{read_npx}
\alias{read_npx}
\title{Read NPX and metadata file together}
\usage{
read_npx(npxFn, metaFn)
}
\arguments{
\item{npxFn}{Path to NPX files}

\item{metaFn}{Path to metadata excel file}
}
\value{
A list with two objects: a \code{\link[tibble]{tibble}} in long format and a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object.
}
\description{
Read one or more NPX files from Olink NPX manager and
one metadata file in excel format.
}
\details{
The metadata excel file shall have column \dQuote{SampleID} and other metadata column that ends with \dQuote{_Factor} (mandatory )or \dQuote{_Numeric} (optional).
If some samples from NPX don't exist in metadata file, it gives a warning.
Only the samples in metadata file are kept.
If some samples from metadata file don't exists in NPX file, it gives an error.
}
\examples{
npxFn <- system.file("extdata", c("20200507_Inflammation_NPX_1.xlsx",
                                  "20200625_Inflammation_NPX_2.xlsx"),
                     package="OlinkR")
metaFn <- system.file("extdata", "Inflammation_Metadata.xlsx", package="OlinkR")
read_npx(npxFn, metaFn)
}
\author{
Ge Tan
}
